requirements.txt
langchain
python-dotenv
langchain-community
sympy
langchain-ollama
llama-cpp-python
huggingface-hub
fastapi
uvicorn